refactor 24.10.14 피드백 반영
