+++ src/main/java/lotto/domain/LottoNumber.java
@@ -4,7 +4,8 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class RandomLottoNumbers {
+public class LottoNumber {
+    private static final String IS_NOT_LOTTO_NUMBER = "로또번호가 1~45가 아닙니다.";
     private static final int LOTTO_MIN_NUM = 1;
     private static final int LOTTO_MAX_NUM = 45;
     private static final List<Integer> randomLottoNumbers = new ArrayList<>();
@@ -15,9 +16,27 @@ public class RandomLottoNumbers {
         }
     }
 
+    private final int lottoNumber;
+
+    public LottoNumber(int lottoNumber) {
+        validateLottoNumbers(lottoNumber);
+        this.lottoNumber = lottoNumber;
+    }
+
     public static List<Integer> generateAutoLottoNumbers() {
         List<Integer> numbers = new ArrayList<>(randomLottoNumbers);
         Collections.shuffle(numbers);
         return numbers.subList(0, 6);
+    }
+
+
+    private void validateLottoNumbers(int lottoNumber) {
+        if (lottoNumber < LOTTO_MIN_NUM || lottoNumber > LOTTO_MAX_NUM) {
+            throw new IllegalArgumentException(IS_NOT_LOTTO_NUMBER);
+        }
+    }
+
+    public int getLottoNumber() {
+        return lottoNumber;
     }
 }
