+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,0 +1,5 @@
+package lotto.domain;
+
+class LottoNumberTest {
+
+}
